Updated brew install instructions to just install libu2f-host.
After actually reading the changes (doh!) I realised that they were for installing the dependencies needed to build libu2f-host.

My original point with submodules was to avoid having to use brew for anything. If the submodules require brew to install tools to build, then we're no better off, and we might as well just give up on submodules and use brew to install the libs themselves!